Changed algorithm in similar_docs to match clusters.
In particular, now iterates over subgraph of similarities
above cutoff to find all docs reachable from target doc.
This has two advantages to the otherwise simpler method
of returning all docs directly within the cutoff of the
target. First, it matches what's returned by the clusters
method, so that we only have one notion consistent
notion of grouping by similarity. Second, it's transitive,
so that for a given cutoff if A and B are in a similar doc
and B and C are in a similar doc set, then A and C will
also be in the set.