Added caching of hierarchy and optional computation of summaries.
Two version of representative_phrases exist in this commit: a
prefetch version and an all SQL version. The all SQL version
is currently used, but that's not a final decision. The prefetch
version has a high initial cost, but can compute many summaries
quickly. The SQL version can compute a single cluster in a reasonable
amount of time, but becomes slower in hierarchies with many clusters.

The cluster() method should probably updated in some way to take
advantage of the cached hierarchy.